+++ test/models/folio/file_test.rb
@@ -343,6 +343,20 @@ class Folio::FileTest < ActiveSupport::TestCase
       end
     end
   end
+
+  test "neutral slug format is used when no headline is present" do
+    with_config(folio_image_metadata_extraction_enabled: false) do
+      image = create(:folio_file_image, headline: nil, slug: nil)
+      assert_match(/\A\d{10}-[0-9a-f]{10}\z/, image.slug)
+    end
+  end
+
+  test "DB has a unique index on folio_files.slug" do
+    indexes = ActiveRecord::Base.connection.indexes(:folio_files)
+    slug_idx = indexes.find { |i| i.columns == ["slug"] }
+    assert slug_idx, "Expected an index on folio_files.slug"
+    assert slug_idx.unique, "Expected slug index to be unique"
+  end
 end
 
 class Folio::FileUrlOrPathTest < ActiveSupport::TestCase